requirements.txt
asgiref==3.2.3
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
django-cors-headers
PyYAML
uritemplate
